return response code and updat test

+++ app/Http/Controllers/CommentsController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Comments;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Validator;
 
 class CommentsController extends Controller
 {
@@ -37,12 +39,23 @@ class CommentsController extends Controller
      */
     public function store(Request $request)
     {
+        $parameters = array('user_id' => $request->input('user_id'),
+            'video_id' => $request->input('video_id'));
+        $validation = Validator::make($parameters, [
+            'user_id' => 'required|integer',
+            'video_id' => 'required|integer',
+        ]);
+        if ($validation->fails()) {
+            return abort(412,'Precondition Failed');
+        }
         $data = $request->input('data');
         $comment = Comments::create(['user_id' => $request->input('user_id'),
             'video_id' => $request->input('user_id'),
             'content' => $data['content']]);
-
-        return json_encode(array('success' => true, 'comment_id' => $comment->comment_id));
+        return response()->json([
+            'success' => true,
+            'comment_id' => $comment->comment_id
+        ], 200);
     }
 
     /**
@@ -53,9 +66,19 @@ class CommentsController extends Controller
      */
     public function show(Request $request)
     {
+        $parameters = array('video_id' => $request->input('video_id'));
+        $validation = Validator::make($parameters, [
+            'video_id' => 'required|integer'
+        ]);
+        if ($validation->fails()) {
+            return abort(412,'Precondition Failed');
+        }
         $video_id = $request->input('video_id');
         $comments = Comments::all()->where('video_id', '=', $video_id);
-        return json_encode(array('success' => true, 'comments' => $comments));
+        return response()->json([
+            'success' => true,
+            'comment' => $comments
+        ], 200);
     }
 
     /**
@@ -78,9 +101,18 @@ class CommentsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $parameters = array('user_id' => $request->input('user_id'),
+            'comment_id' => $id);
+        $validation = Validator::make($parameters, [
+            'user_id' => 'required|integer',
+            'comment_id' => 'required|integer'
+        ]);
+        if ($validation->fails()) {
+            return abort(412,'Precondition Failed');
+        }
         $comment = Comments::find($id);
         if (!$comment) {
-            return json_encode(array('success' => false, 'message' => 'Comment not found'));
+            return abort(404, 'Resource Not Found');
         }
         if ($comment->user_id == $request->input('user_id')) {
             $data = $request->input('data');
@@ -88,9 +120,11 @@ class CommentsController extends Controller
             DB::table('comments')
                 ->where('comment_id', $id)
                 ->update(['content' => $newContent]);
-            return json_encode(array('success' => true));
+            return response()->json([
+                'success' => true
+            ], 200);
         }
-        return json_encode(array('success' => false, 'message' => 'You do not have the permission to edit this comment'));
+        return abort(403, 'Permission Denied');
     }
 
     /**
@@ -98,18 +132,31 @@ class CommentsController extends Controller
      *
      * @param  int $id
      * @return \Illuminate\Http\Response
+     * @throws \Exception
      */
     public function destroy(Request $request, $id)
     {
+        $parameters = array('user_id' => $request->input('user_id'),
+            'comment_id' => $id);
+        $validation = Validator::make($parameters, [
+            'user_id' => 'required|integer',
+            'comment_id' => 'required|integer'
+        ]);
+        if ($validation->fails()) {
+            return abort(412,'Precondition Failed');
+        }
         $comment = Comments::find($id);
         if ($comment) {
             if ($comment->user_id != $request->input('user_id')) {
-                return json_encode(array('success' => false, 'message' => 'You do not have the permission to delete this comment'));
+                return abort(403, 'Permission Denied');
             }
             $comment->delete();
-            return json_encode(array('success' => true));
+            return response()->json([
+                'success' => true
+            ], 200);
+//            return json_encode(array('success' => true));
         }
-        return json_encode(array('success' => false, 'message' => 'Comment not found'));
+        return abort(404, 'Resource Not Found');
 
 
     }
